% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{get_habitat_3857}
\alias{get_habitat_3857}
\title{habitat raster}
\usage{
get_habitat_3857()
}
\value{
the habitat \code{\link[terra:SpatRaster-class]{terra::SpatRaster}} (EPSG:3857) describing suitability for beaver settlement/dispersal, to be shipped with the package
}
\description{
raster for the beaver app, stored in beaverAppData/data/extdata because large
}
